feat: users need to be able to revoke a token without knowing it
Sometimes, users would need to for instance a revoke a user's token
just by knowing the user's ID(which they can use to generate a
token). So I replaced random generation of tokens with the hmac of
the key or passed key. All this relies on the fact that our secret
key is never stolen 😭